requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
sqlalchemy==2.0.48
alembic==1.16.5
psycopg2-binary==2.9.11
asyncpg==0.31.0
pydantic==2.12.5
python-jose==3.5.0
python-dotenv==1.2.1
passlib==1.7.4
bcrypt==5.0.0
cryptography==46.0.5
python-multipart==0.0.20
